remove the print used for debug in random_crop
