+++ paddlex/cv/models/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 from .unet import UNet
-from .classifier import BaseClassifier
+from .classifier import ResNet18, ResNet34, ResNet50, ResNet101, ResNet152
